+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Dnsmasq Local Cookbook CHANGELOG
 
 Unreleased
 ----------
+- Fix custom config properties/attributes under Chef 11
 
 v0.2.0 (2016-05-18)
 -------------------